+++ SDK/backendlessAPI/Classes/weborb/Reader/V3ObjectReader.m
@@ -174,7 +174,7 @@
 }
 
 -(int)getPropertyCount {
-    return props.count;
+    return (int)props.count;
 }
 
 -(NSString *)getProperty:(int)index {
